Mistress::Env::Config: huge reworking
The previous design was making impossible to use something else than the
filesystem to get a configuration. Now, MEC is much more generic, and
Mistress::Env::Config::Base has become Mistress::Env::Config::File because all
filesystem-specific stuff is inside.

From there, it will be easy to define a Mistress::Env::Config::Hash (that does
not depend on any file but works in memory only) for use in tests.